+++ packages/astro/src/cli/load-settings.ts
@@ -1,8 +1,9 @@
 /* eslint-disable no-console */
-import fs from 'fs';
 import * as colors from 'kleur/colors';
+import fs from 'node:fs';
 import type { Arguments as Flags } from 'yargs-parser';
 import { ZodError } from 'zod';
+import type { AstroInlineConfig } from '../@types/astro.js';
 import {
 	createSettings,
 	resolveConfig,
@@ -14,7 +15,6 @@ import { error, type LogOptions } from '../core/logger/core.js';
 import { formatConfigErrorMessage, formatErrorMessage } from '../core/messages.js';
 import * as event from '../events/index.js';
 import { eventConfigError, telemetry } from '../events/index.js';
-import type { AstroInlineConfig } from '../@types/astro.js';
 
 interface LoadSettingsOptions {
 	cmd: string;
@@ -68,8 +68,6 @@ export function flagsToAstroInlineConfig(flags: Flags): AstroInlineConfig {
 		},
 		experimental: {
 			assets: typeof flags.experimentalAssets === 'boolean' ? flags.experimentalAssets : undefined,
-			redirects:
-				typeof flags.experimentalRedirects === 'boolean' ? flags.experimentalRedirects : undefined,
 		},
 	};
 }