+++ packages/astro/src/core/config/schema.ts
@@ -45,7 +45,7 @@ const ASTRO_CONFIG_DEFAULTS = {
 	redirects: {},
 	experimental: {
 		assets: false,
-		redirects: false,
+		viewTransitions: false,
 	},
 } satisfies AstroUserConfig & { server: { open: boolean } };
 
@@ -233,7 +233,10 @@ export const AstroConfigSchema = z.object({
 	experimental: z
 		.object({
 			assets: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.experimental.assets),
-			redirects: z.boolean().optional().default(ASTRO_CONFIG_DEFAULTS.experimental.redirects),
+			viewTransitions: z
+				.boolean()
+				.optional()
+				.default(ASTRO_CONFIG_DEFAULTS.experimental.viewTransitions),
 		})
 		.passthrough()
 		.refine(