+++ packages/astro/src/runtime/server/index.ts
@@ -39,6 +39,7 @@ export type {
 	ComponentSlots,
 	RenderInstruction,
 } from './render/index.js';
+export { renderTransition } from './transition.js';
 
 import { markHTMLString } from './escape.js';
 import { addAttribute, Renderer } from './render/index.js';